bin/death_sentence_ble-main 2/death_sentence/agents/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
jinja2>=3.1.4
openai>=1.51.0
python-dotenv>=1.0.1

